0.14  3 September 2006 ---------------------- Download from http://rocket-gwt.googlecode.com/svn/tags/0.14/Rocket-0.14.zip
MOD Upgraded to GWT release 1.1.10, Changes listed below
MOD return type of add( Widget) and insert( Widget) from boolean to void
MOD argument of Timer.scheduleRepeating() from long to int
MOD return type of DOM.eventGetCharCode() from char to int
MOD Updated *.launch *.cmd to use new library release
DEL Deleted DeckPanel
MOD CssUnitScalingFileServlet query parameter names are no longer set via init-parameters but use values from **.server.BrowserConstants.

+++ Rocket/src/rocket/test/widget/life/client/LifeTest.java
@@ -115,7 +115,7 @@ public class LifeTest implements EntryPoint {
             this.setCounter(this.getCounter() + 1);
         }
 
-        public void scheduleRepeating(long interval) {
+        public void scheduleRepeating(int interval) {
             setStartTime(System.currentTimeMillis());
             super.scheduleRepeating(interval);
         }
